Failsafe tests to be sure that we are deleting an existing unpacked sources directory. Thanks to Petr "Qaxi" Klíma for the report and patch.

+++ BLFS/libs/scripts.xsl
@@ -155,7 +155,7 @@ cd xc&#xA;</xsl:text>
         <xsl:text>
 if [[ -e unpacked ]] ; then
   UNPACKDIR=`head -n1 unpacked | sed 's@^./@@;s@/.*@@'`
-  rm -rf $UNPACKDIR
+  [[ -n $UNPACKDIR ]] && [[ -d $UNPACKDIR ]] && rm -rf $UNPACKDIR
 fi
 tar -xvf $PACKAGE > unpacked
 UNPACKDIR=`head -n1 unpacked | sed 's@^./@@;s@/.*@@'`
